LABTSP/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
alembic==1.12.1
bcrypt==4.0.1
python-dotenv==1.0.0
networkx==3.2.1
websockets==12.0
celery==5.3.6
redislite==6.2.830
click==8.1.7 